Renamed packages to better reflect the reality of MVC
