+++ backend-treasurehunt/src/main/java/com/geoschnitzel/treasurehunt/backend/schema/Target.java
@@ -1,28 +1,36 @@
-package com.geoschnitzel.treasurehunt.backend.model;
+package com.geoschnitzel.treasurehunt.backend.schema;
 
+import java.util.List;
+
+import javax.persistence.CascadeType;
+import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.OneToMany;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * A single step in a treasure hunt
+ */
 @Entity
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
-public class User {
+public class Target {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private Long id;
 
-    //TODO Google ID oder so
+    @Embedded
+    private Area area;
 
-    private String displayName;
-
-    private String email;
+    @OneToMany(cascade = CascadeType.ALL)
+    private List<Hint> hints;
 
 }
