+++ backend-treasurehunt/src/test/java/com/geoschnitzel/treasurehunt/backend/MessageRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.geoschnitzel.treasurehunt.backend;
 
-import com.geoschnitzel.treasurehunt.backend.model.Message;
-import com.geoschnitzel.treasurehunt.backend.repository.MessageRepository;
+import com.geoschnitzel.treasurehunt.backend.model.MessageRepository;
+import com.geoschnitzel.treasurehunt.backend.schema.Message;
 
 import org.junit.Before;
 import org.junit.Test;
